Add null check, tmpDir may not have been initialized

+++ RCaller/src/main/java/com/github/rcaller/rstuff/RCaller.java
@@ -441,7 +441,9 @@ public class RCaller {
             process.destroy();
             stopStreamConsumers();
             process = null;
-            deleteDirectory(new File(tmpDir));
+            if (tmpDir != null) {
+                deleteDirectory(new File(tmpDir));
+            }
         }
     }
